+++ DoyProject/Doy/Doy_Backend/src/main/java/com/pingfloyd/doy/controllers/RestaurantController.java
@@ -3,7 +3,12 @@ package com.pingfloyd.doy.controllers;
 import com.pingfloyd.doy.dto.DtoRestaurant;
 import com.pingfloyd.doy.dto.DtoRestaurantIU;
 import com.pingfloyd.doy.dto.RestaurantRequest;
+import com.pingfloyd.doy.entities.UserRoles;
+import com.pingfloyd.doy.enums.RestaurantCategory;
+import com.pingfloyd.doy.exception.UnauthorizedRequestException;
+import com.pingfloyd.doy.jwt.JwtService;
 import com.pingfloyd.doy.services.RestaurantService;
+import com.pingfloyd.doy.services.UserService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -19,6 +24,10 @@ public class RestaurantController implements IRestaurantController {
 
     @Autowired
     private RestaurantService restaurantService;
+    @Autowired
+    private JwtService jwtService;
+    @Autowired
+    private UserService userService;
 
     @Override
     @GetMapping("/get/{id}")
@@ -27,31 +36,27 @@ public class RestaurantController implements IRestaurantController {
     }
 
     @Override
-    @PostMapping("/post")
-    //@PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<DtoRestaurant> postRestaurant(@RequestBody @Valid DtoRestaurantIU dtoRestaurantIU) {
-        return ResponseEntity.ok(restaurantService.postRestaurant(dtoRestaurantIU));
-    }
-
-    @Override
     @PutMapping("/update/{id}")
-    //@PreAuthorize("hasAuthority('RESTAURANT_OWNER') or hasAuthority('ADMIN')")
     public ResponseEntity<DtoRestaurant> updateRestaurant(@PathVariable(name = "id") Long id, @RequestBody @Valid DtoRestaurantIU dtoRestaurantIU) {
-        return ResponseEntity.ok(restaurantService.updateRestaurant(id, dtoRestaurantIU));
+        if ((jwtService.checkIfUserRole(UserRoles.RESTAURANT_OWNER) &&
+        userService.checkIfSameUserFromToken(id)) || jwtService.checkIfUserRole(UserRoles.ADMIN))
+            return ResponseEntity.ok(restaurantService.updateRestaurant(id, dtoRestaurantIU));
+        throw new UnauthorizedRequestException();
+
     }
 
     @Override
     @DeleteMapping("/delete/{id}")
-    //@PreAuthorize("hasAuthority('ADMIN')") // only admins can delete
     public ResponseEntity<DtoRestaurant> deleteRestaurant(@PathVariable(name = "id") Long id) {
+        if (!jwtService.checkIfUserRole(UserRoles.ADMIN)) throw new UnauthorizedRequestException();
         return ResponseEntity.ok(restaurantService.deleteRestaurant(id));
     }
 
 
     @Override
     @GetMapping("/get-all")
-    //@PreAuthorize("hasAuthority('ADMIN')")
     public ResponseEntity<List<DtoRestaurant>> getAllRestaurants() {
+        if (!jwtService.checkIfUserRole(UserRoles.ADMIN)) throw new UnauthorizedRequestException();
         return ResponseEntity.ok(restaurantService.gelAllRestaurants());
     }
 
@@ -65,6 +70,11 @@ public class RestaurantController implements IRestaurantController {
     public ResponseEntity<Boolean> setFavoriteRestaurant(@PathVariable(name = "restaurantId") Long restaurantId) {
         String username = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
         return ResponseEntity.ok(restaurantService.setFavoriteRestaurant(username, restaurantId));
+    }
+
+    @GetMapping("/get-types")
+    public ResponseEntity<RestaurantCategory[]> GetCategories(){
+        return ResponseEntity.ok(restaurantService.GetCategories());
     }
 
     @GetMapping("/favorite")
